move regexp integration tests
