+++ sql/src/test/java/io/crate/integrationtests/TransportSQLActionTest.java
@@ -2420,31 +2420,6 @@ public class TransportSQLActionTest extends SQLTransportIntegrationTest {
     }
 
     @Test
-    public void testRegexpMatchOperator() throws Exception {
-        this.setup.setUpLocations();
-        ensureGreen();
-        refresh();
-        execute("select distinct name from locations where name ~ '[A-Z][a-z0-9]+' order by name");
-        assertThat(response.rowCount(), is(5L));
-        assertThat((String) response.rows()[0][0], is("Aldebaran"));
-        assertThat((String) response.rows()[1][0], is("Algol"));
-        assertThat((String) response.rows()[2][0], is("Altair"));
-        assertThat((String) response.rows()[3][0], is("Argabuthon"));
-        assertThat((String) response.rows()[4][0], is("Bartledan"));
-
-        execute("select name from locations where name !~ '[A-Z][a-z0-9]+' order by name");
-        assertThat(response.rowCount(), is(8L));
-        assertThat((String) response.rows()[0][0], is(""));
-        assertThat((String) response.rows()[1][0], is("Allosimanius Syneca"));
-        assertThat((String) response.rows()[2][0], is("Alpha Centauri"));
-        assertThat((String) response.rows()[3][0], is("Arkintoofle Minor"));
-        assertThat((String) response.rows()[4][0], is("Galactic Sector QQ7 Active J Gamma"));
-        assertThat((String) response.rows()[5][0], is("North West Ripple"));
-        assertThat((String) response.rows()[6][0], is("Outer Eastern Rim"));
-        assertThat(response.rows()[7][0], is(nullValue()));
-    }
-
-    @Test
     public void testWhereColumnEqColumnAndFunctionEqFunction() throws Exception {
         this.setup.setUpLocations();
         ensureGreen();
